requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16

drf-nested-routers
django-extensions
django-cors-headers==3.13.0
django-countries==7.4.2
django-phonenumber-field==7.0.0
phonenumbers==8.13.1
Pillow>=8.2.0,<8.3.0

amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backoff==2.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
geographiclib==2.0
geopy==2.3.0
idna==3.4
ipapi==1.0.4
ipify==1.0.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.32
pyparsing==3.0.9
pytz==2022.6
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1